Add: IsAllowedMethod

+++ srcs/http/response/http_handle_method.cpp
@@ -9,6 +9,7 @@
 #include <iostream>
 #include <sstream>
 #include <unistd.h> // access
+#include <algorithm> // std::find
 
 namespace {
 
@@ -42,19 +43,23 @@ std::string ReadFile(const std::string &file_path) {
 
 namespace http {
 
-StatusCode
-HttpResponse::MethodHandler(const std::string& path, const std::string& method, const std::list<std::string>& allow_method, const std::string& request_body_message, std::string& response_body_message) {
-	StatusCode  status_code(OK);
-	(void) allow_method;
-	// 引数: path, allow_method, method, request_body_message, response_body_message
-	if (method == GET) {
+StatusCode HttpResponse::MethodHandler(
+	const std::string            &path,
+	const std::string            &method,
+	const std::list<std::string> &allow_method,
+	const std::string            &request_body_message,
+	std::string                  &response_body_message
+) {
+	StatusCode status_code(OK);
+	bool       is_allow_method = IsAllowedMethod(method, allow_method);
+	if (is_allow_method && method == GET) {
 		status_code = GetHandler(path, response_body_message);
-	} else if (method == POST) {
+	} else if (is_allow_method && method == POST) {
 		status_code = PostHandler(path, request_body_message, response_body_message);
-	} else if (method == DELETE) {
+	} else if (is_allow_method && method == DELETE) {
 		status_code = DeleteHandler(path, response_body_message);
 	} else {
-		status_code      = StatusCode(NOT_IMPLEMENTED);
+		status_code           = StatusCode(NOT_IMPLEMENTED);
 		response_body_message = CreateDefaultBodyMessageFormat(status_code);
 		throw HttpException("Error: Not Implemented", status_code);
 	}
@@ -191,6 +196,21 @@ Stat HttpResponse::TryStat(const std::string &path, std::string &response_body_m
 	}
 	Stat info(stat_buf);
 	return info;
+}
+
+bool HttpResponse::IsAllowedMethod(
+	const std::string &method, const std::list<std::string> &allow_method
+) {
+	if (allow_method.empty()) {
+		// allow_methodがない場合はwebservが許可したメソッドのみ許可する（GETのみ）
+		return std::find(
+				   DEFAULT_ALLOWED_METHODS,
+				   DEFAULT_ALLOWED_METHODS + DEFAULT_ALLOWED_METHODS_SIZE,
+				   method
+			   ) != DEFAULT_ALLOWED_METHODS + DEFAULT_ALLOWED_METHODS_SIZE;
+	} else {
+		return std::find(allow_method.begin(), allow_method.end(), method) != allow_method.end();
+	}
 }
 
 } // namespace http
